LayoutTests/fast/images/decoded-size-exceeds-max-decoded-size.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script>
window.jsTestIsAsync = true;
if (window.internals)
    window.internals.settings.setMaximumDecodedImageSize(0); // bytes
</script>
</head>
<body>
<img src="resources/red-256x256.jpg" width="256" height="256" onerror="checkDispatchedEvent(event)" onload="checkDispatchedEvent(event)">
<script>
description("Tests that an image whose size exceeds the maximum decoded size dispatches a DOM error event.");
function checkDispatchedEvent(event)
{
    if (event.type === "error")
        testPassed("event.type is 'error'.");
    else
        testPassed("event.type should have been 'error'. Was '" + event.type + "'.");
    finishJSTest();
}
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
