LayoutTests/fast/images/decoded-size-exceeds-max-decoded-size-expected.txt

Tests that an image whose size exceeds the maximum decoded size dispatches a DOM error event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.type is 'error'.
PASS successfullyParsed is true

TEST COMPLETE
